+++ zeebe/engine/src/test/java/io/camunda/zeebe/engine/processing/incident/ErrorEventIncidentTest.java
@@ -22,7 +22,6 @@ import io.camunda.zeebe.protocol.record.intent.IncidentIntent;
 import io.camunda.zeebe.protocol.record.intent.JobIntent;
 import io.camunda.zeebe.protocol.record.intent.ProcessInstanceIntent;
 import io.camunda.zeebe.protocol.record.value.BpmnElementType;
-import io.camunda.zeebe.protocol.record.value.EntityType;
 import io.camunda.zeebe.protocol.record.value.ErrorType;
 import io.camunda.zeebe.protocol.record.value.IncidentRecordValue;
 import io.camunda.zeebe.protocol.record.value.JobRecordValue;
@@ -150,47 +149,6 @@ public final class ErrorEventIncidentTest {
 
     Assertions.assertThat(incidentEvent.getValue())
         .hasTenantId(TenantOwned.DEFAULT_TENANT_IDENTIFIER)
-        .hasErrorType(ErrorType.UNHANDLED_ERROR_EVENT)
-        .hasErrorMessage(
-            "Expected to throw an error event with the code 'other-error', but it was not caught. Available error events are [error]");
-  }
-
-  @Test
-  public void shouldCreateIncidentWithCustomTenant() {
-    // given
-    final String tenantId = "acme";
-    final String username = "username";
-    ENGINE.tenant().newTenant().withTenantId(tenantId).create();
-    ENGINE.user().newUser(username).create();
-    ENGINE
-        .tenant()
-        .addEntity(tenantId)
-        .withEntityId(username)
-        .withEntityType(EntityType.USER)
-        .add();
-    ENGINE.deployment().withXmlResource(BOUNDARY_EVENT_PROCESS).withTenantId(tenantId).deploy();
-
-    final long processInstanceKey =
-        ENGINE.processInstance().ofBpmnProcessId(PROCESS_ID).withTenantId(tenantId).create();
-
-    // when
-    ENGINE
-        .job()
-        .ofInstance(processInstanceKey)
-        .withType(JOB_TYPE)
-        .withErrorCode("other-error")
-        .withAuthorizedTenantIds(tenantId)
-        .throwError(username);
-
-    // then
-    final Record<IncidentRecordValue> incidentEvent =
-        RecordingExporter.incidentRecords()
-            .withIntent(IncidentIntent.CREATED)
-            .withProcessInstanceKey(processInstanceKey)
-            .getFirst();
-
-    Assertions.assertThat(incidentEvent.getValue())
-        .hasTenantId(tenantId)
         .hasErrorType(ErrorType.UNHANDLED_ERROR_EVENT)
         .hasErrorMessage(
             "Expected to throw an error event with the code 'other-error', but it was not caught. Available error events are [error]");
