+++ zeebe/engine/src/test/java/io/camunda/zeebe/engine/processing/incident/JobFailIncidentTest.java
@@ -24,16 +24,13 @@ import io.camunda.zeebe.protocol.record.intent.IncidentIntent;
 import io.camunda.zeebe.protocol.record.intent.JobBatchIntent;
 import io.camunda.zeebe.protocol.record.intent.JobIntent;
 import io.camunda.zeebe.protocol.record.intent.ProcessInstanceIntent;
-import io.camunda.zeebe.protocol.record.value.EntityType;
 import io.camunda.zeebe.protocol.record.value.ErrorType;
 import io.camunda.zeebe.protocol.record.value.IncidentRecordValue;
 import io.camunda.zeebe.protocol.record.value.JobRecordValue;
 import io.camunda.zeebe.protocol.record.value.ProcessInstanceRecordValue;
-import io.camunda.zeebe.test.util.Strings;
 import io.camunda.zeebe.test.util.collection.Maps;
 import io.camunda.zeebe.test.util.record.RecordingExporter;
 import io.camunda.zeebe.test.util.record.RecordingExporterTestWatcher;
-import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
@@ -148,37 +145,6 @@ public final class JobFailIncidentTest {
         .hasElementId("failingTask")
         .hasElementInstanceKey(activityEvent.getKey())
         .hasVariableScopeKey(activityEvent.getKey());
-  }
-
-  @Test
-  public void shouldCreateIncidentIfJobHasNoRetriesLeftWithCustomTenant() {
-    // given
-    final String processId = "test-process";
-    final String tenantId = Strings.newRandomValidIdentityId();
-    final String username = Strings.newRandomValidIdentityId();
-    ENGINE.tenant().newTenant().withTenantId(tenantId).create();
-    ENGINE.user().newUser(username).create();
-    ENGINE
-        .tenant()
-        .addEntity(tenantId)
-        .withEntityId(username)
-        .withEntityType(EntityType.USER)
-        .add();
-    final Record<JobRecordValue> jobRecord =
-        ENGINE.createJob(JOB_TYPE, processId, Collections.emptyMap(), tenantId);
-    final long piKey = jobRecord.getValue().getProcessInstanceKey();
-
-    // when
-    ENGINE.job().withType(JOB_TYPE).withRetries(0).ofInstance(piKey).fail(username);
-
-    // then
-    final Record<IncidentRecordValue> incidentEvent =
-        RecordingExporter.incidentRecords()
-            .withIntent(IncidentIntent.CREATED)
-            .withProcessInstanceKey(piKey)
-            .getFirst();
-
-    assertThat(incidentEvent.getValue()).hasTenantId(tenantId);
   }
 
   @Test
@@ -355,73 +321,6 @@ public final class JobFailIncidentTest {
             tuple(RecordType.EVENT, ValueType.JOB, JobIntent.FAILED),
             tuple(RecordType.COMMAND, ValueType.JOB, JobIntent.UPDATE_RETRIES),
             tuple(RecordType.EVENT, ValueType.JOB, JobIntent.RETRIES_UPDATED));
-  }
-
-  @Test
-  public void shouldResolveIncidentWithCustomTenant() {
-    // given
-    final String processId = "test-process";
-    final String tenantId = Strings.newRandomValidIdentityId();
-    final String username = Strings.newRandomValidIdentityId();
-    ENGINE.tenant().newTenant().withTenantId(tenantId).create();
-    ENGINE.user().newUser(username).create();
-    ENGINE
-        .tenant()
-        .addEntity(tenantId)
-        .withEntityId(username)
-        .withEntityType(EntityType.USER)
-        .add();
-    final Record<JobRecordValue> jobRecord =
-        ENGINE.createJob(JOB_TYPE, processId, Collections.emptyMap(), tenantId);
-    final long piKey = jobRecord.getValue().getProcessInstanceKey();
-    ENGINE.job().withType(JOB_TYPE).withRetries(0).ofInstance(piKey).fail(username);
-
-    // when
-    ENGINE.job().ofInstance(piKey).withType(JOB_TYPE).withRetries(1).updateRetries(username);
-
-    final Record<IncidentRecordValue> resolvedIncident =
-        ENGINE.incident().ofInstance(piKey).resolve(username);
-
-    // then
-    assertThat(resolvedIncident.getValue()).hasTenantId(tenantId);
-    assertThat(resolvedIncident).hasIntent(IncidentIntent.RESOLVED);
-  }
-
-  @Test
-  public void shouldRejectResolvingIncidentWithUnauthorizedTenant() {
-    // given
-    final String processId = "test-process";
-    final String tenantId = Strings.newRandomValidIdentityId();
-    final String unauthorizedTenantId = Strings.newRandomValidIdentityId();
-    final String authorizedUser = Strings.newRandomValidIdentityId();
-    final String unauthorizedUser = Strings.newRandomValidIdentityId();
-    ENGINE.tenant().newTenant().withTenantId(tenantId).create();
-    ENGINE.tenant().newTenant().withTenantId(unauthorizedTenantId).create();
-    ENGINE.user().newUser(authorizedUser).create();
-    ENGINE
-        .tenant()
-        .addEntity(tenantId)
-        .withEntityId(authorizedUser)
-        .withEntityType(EntityType.USER)
-        .add();
-    ENGINE.user().newUser(unauthorizedUser).create();
-    ENGINE
-        .tenant()
-        .addEntity(unauthorizedTenantId)
-        .withEntityId(unauthorizedUser)
-        .withEntityType(EntityType.USER)
-        .add();
-    final Record<JobRecordValue> jobRecord =
-        ENGINE.createJob(JOB_TYPE, processId, Collections.emptyMap(), tenantId);
-    final long piKey = jobRecord.getValue().getProcessInstanceKey();
-    ENGINE.job().withType(JOB_TYPE).withRetries(0).ofInstance(piKey).fail(authorizedUser);
-
-    // when
-    final Record<IncidentRecordValue> resolvedIncident =
-        ENGINE.incident().ofInstance(piKey).expectRejection().resolve(unauthorizedUser);
-
-    // then
-    assertThat(resolvedIncident).hasRejectionType(RejectionType.NOT_FOUND);
   }
 
   @Test
