+++ zeebe/engine/src/test/java/io/camunda/zeebe/engine/processing/job/JobUpdateRetriesTest.java
@@ -23,7 +23,6 @@ import io.camunda.zeebe.protocol.record.value.JobRecordValue;
 import io.camunda.zeebe.protocol.record.value.PermissionType;
 import io.camunda.zeebe.test.util.Strings;
 import io.camunda.zeebe.test.util.record.RecordingExporterTestWatcher;
-import java.util.Collections;
 import java.util.UUID;
 import org.junit.Before;
 import org.junit.BeforeClass;
@@ -47,9 +46,7 @@ public final class JobUpdateRetriesTest {
   public static void setUp() {
     tenantId = UUID.randomUUID().toString();
     username = UUID.randomUUID().toString();
-    final var user = ENGINE.user().newUser(username).create().getValue();
-    final var userKey = user.getUserKey();
-    final var username = user.getUsername();
+    ENGINE.user().newUser(username).create().getValue();
     ENGINE.tenant().newTenant().withTenantId(tenantId).create().getValue().getTenantKey();
     ENGINE
         .tenant()
@@ -192,44 +189,5 @@ public final class JobUpdateRetriesTest {
 
     // then
     Assertions.assertThat(jobRecord).hasRejectionType(RejectionType.INVALID_ARGUMENT);
-  }
-
-  @Test
-  public void shouldUpdateRetriesForCustomTenant() {
-    // given
-    ENGINE.createJob(jobType, PROCESS_ID, Collections.emptyMap(), tenantId);
-    final Record<JobBatchRecordValue> batchRecord =
-        ENGINE.jobs().withType(jobType).withTenantId(tenantId).activate(username);
-    final long jobKey = batchRecord.getValue().getJobKeys().get(0);
-
-    // when
-    final Record<JobRecordValue> updatedRecord =
-        ENGINE.job().withKey(jobKey).withRetries(NEW_RETRIES).updateRetries(username);
-
-    // then
-    Assertions.assertThat(updatedRecord.getValue()).hasTenantId(tenantId);
-  }
-
-  @Test
-  public void shouldRejectUpdateRetriesIfTenantIsUnauthorized() {
-    // given
-    final String falseTenantId = "foo";
-    ENGINE.createJob(jobType, PROCESS_ID, Collections.emptyMap(), tenantId);
-    final Record<JobBatchRecordValue> batchRecord =
-        ENGINE.jobs().withType(jobType).withTenantId(tenantId).activate(username);
-    final long jobKey = batchRecord.getValue().getJobKeys().get(0);
-
-    // when
-    final Record<JobRecordValue> jobRecord =
-        ENGINE
-            .job()
-            .withKey(jobKey)
-            .withRetries(NEW_RETRIES)
-            .withAuthorizedTenantIds(falseTenantId)
-            .expectRejection()
-            .updateRetries();
-
-    // then
-    Assertions.assertThat(jobRecord).hasRejectionType(RejectionType.NOT_FOUND);
   }
 }
